+++ src/Modules/Stakeholders/Explorer.Stakeholders.API/Dtos/ProblemDto.cs
@@ -1,10 +1,12 @@
-﻿namespace Explorer.Tours.API.Dtos;
+namespace Explorer.Stakeholders.API.Dtos;
+
 public enum ProblemCategory
 {
     Safety,
     Maintenance,
     Other
 }
+
 public class ProblemDto
 {
     public long Id { get; set; }
@@ -14,6 +16,4 @@ public class ProblemDto
     public string Description { get; set; }
     public DateTime CreationTime { get; set; }
     public ProblemCategory Category { get; set; }
-
 }
-
