+++ src/Modules/Stakeholders/Explorer.Stakeholders.Core/Domain/RepositoryInterfaces/IProblemRepository.cs
@@ -1,11 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using Explorer.BuildingBlocks.Core.UseCases;
 
-namespace Explorer.Tours.Core.Domain.RepositoryInterfaces;
+namespace Explorer.Stakeholders.Core.Domain.RepositoryInterfaces;
 
 public interface IProblemRepository
 {
