+++ src/Modules/Stakeholders/Explorer.Stakeholders.Infrastructure/StakeholdersStartup.cs
@@ -34,6 +34,7 @@ public static class StakeholdersStartup
         services.AddScoped<IRatingsService, RatingsService>();
         services.AddScoped<IUserManagementService, UserManagementService>();
         services.AddScoped<IUserLocationService, UserLocationService>();
+        services.AddScoped<IProblemService, ProblemService>();
     }
 
     private static void SetupInfrastructure(IServiceCollection services)
@@ -48,6 +49,7 @@ public static class StakeholdersStartup
 
         services.AddScoped<IRatingRepository, RatingDbRepository>();
         services.AddScoped<IUserLocationRepository, UserLocationDbRepository>();
+        services.AddScoped<IProblemRepository, ProblemDbRepository>();
 
 
         var dataSourceBuilder = new NpgsqlDataSourceBuilder(DbConnectionStringBuilder.Build("stakeholders"));
