+++ src/Modules/Stakeholders/Explorer.Stakeholders.Tests/TestData/e-problem.sql
@@ -1,11 +1,11 @@
-﻿INSERT INTO tours."Problem"(
+INSERT INTO stakeholders."Problems"(
     "Id", "TourId", "CreatorId", "Priority", "Description", "CreationTime", "Category")
 VALUES (-1, -1, -21, 3, 'Bad organization', '2024-01-15T10:30:00Z', 1);
 
-INSERT INTO tours."Problem"(
+INSERT INTO stakeholders."Problems"(
     "Id", "TourId", "CreatorId", "Priority", "Description", "CreationTime", "Category")
 VALUES (-2, -2, -22, 5, 'The tour was late', '2024-02-20T14:15:00Z', 2);
 
-INSERT INTO tours."Problem"(
+INSERT INTO stakeholders."Problems"(
     "Id", "TourId", "CreatorId", "Priority", "Description", "CreationTime", "Category")
 VALUES (-3, -1, -23, 2, 'Guide was not polite', '2024-03-10T09:45:00Z', 0);
