+++ src/Modules/Tours/Explorer.Tours.Core/Mappers/ToursProfile.cs
@@ -9,7 +9,6 @@ public class ToursProfile : Profile
     public ToursProfile()
     {
         CreateMap<EquipmentDto, Equipment>().ReverseMap();
-        CreateMap<ProblemDto, Problem>().ReverseMap();
         CreateMap<TourDto, Tour>().ReverseMap();
         CreateMap<FacilityDto, Facility>().ReverseMap();
         CreateMap<MeetUpDto, MeetUp>().ReverseMap();
