+++ src/Modules/Tours/Explorer.Tours.Infrastructure/Database/ToursContext.cs
@@ -6,7 +6,6 @@ namespace Explorer.Tours.Infrastructure.Database;
 public class ToursContext : DbContext
 {
     public DbSet<Equipment> Equipment { get; set; }
-    public DbSet<Problem> Problem { get; set; }
     public DbSet<Tour> Tour { get; set; }
     public DbSet<Facility> Facility { get; set; }
     public DbSet<MeetUp> MeetUp { get; set; }
